requirements.txt
Django>=5.1,<6.0
dj-database-url>=2.2
djangorestframework>=3.15
drf-spectacular>=0.27
pgvector>=0.3
psycopg[binary]>=3.2
